Start of Unification for `Sensor`, `Actuator`, and `System` into a Single Class (#2451)
