+++ hardware_interface/CMakeLists.txt
@@ -29,11 +29,9 @@ foreach(Dependency IN ITEMS ${THIS_PACKAGE_INCLUDE_DEPENDS})
 endforeach()
 
 add_library(hardware_interface SHARED
-  src/actuator.cpp
   src/component_parser.cpp
   src/resource_manager.cpp
-  src/sensor.cpp
-  src/system.cpp
+  src/hardware_component.cpp
   src/lexical_casts.cpp
 )
 target_compile_features(hardware_interface PUBLIC cxx_std_17)
