+++ hardware_interface/include/hardware_interface/sensor.hpp
@@ -15,96 +15,11 @@
 #ifndef HARDWARE_INTERFACE__SENSOR_HPP_
 #define HARDWARE_INTERFACE__SENSOR_HPP_
 
-#include <memory>
-#include <string>
-#include <vector>
-
-#include "hardware_interface/handle.hpp"
-#include "hardware_interface/hardware_info.hpp"
+#include "hardware_interface/hardware_component.hpp"
 #include "hardware_interface/sensor_interface.hpp"
-#include "hardware_interface/types/hardware_interface_return_values.hpp"
-#include "hardware_interface/types/statistics_types.hpp"
-#include "rclcpp/duration.hpp"
-#include "rclcpp/logger.hpp"
-#include "rclcpp/node_interfaces/node_clock_interface.hpp"
-#include "rclcpp/time.hpp"
-#include "rclcpp_lifecycle/state.hpp"
 
 namespace hardware_interface
 {
-class SensorInterface;
-
-class Sensor final
-{
-public:
-  Sensor() = default;
-
-  explicit Sensor(std::unique_ptr<SensorInterface> impl);
-
-  explicit Sensor(Sensor && other) noexcept;
-
-  ~Sensor() = default;
-
-  Sensor(const Sensor & other) = delete;
-
-  Sensor & operator=(const Sensor & other) = delete;
-
-  Sensor & operator=(Sensor && other) = delete;
-
-  [[deprecated(
-    "Replaced by const rclcpp_lifecycle::State & initialize(const "
-    "hardware_interface::HardwareComponentParams & params).")]]
-  const rclcpp_lifecycle::State & initialize(
-    const HardwareInfo & sensor_info, rclcpp::Logger logger,
-    rclcpp::node_interfaces::NodeClockInterface::SharedPtr clock_interface);
-
-  [[deprecated(
-    "Replaced by const rclcpp_lifecycle::State & initialize(const "
-    "hardware_interface::HardwareComponentParams & params).")]]
-  const rclcpp_lifecycle::State & initialize(
-    const HardwareInfo & sensor_info, rclcpp::Logger logger, rclcpp::Clock::SharedPtr clock);
-
-  const rclcpp_lifecycle::State & initialize(
-    const hardware_interface::HardwareComponentParams & params);
-
-  const rclcpp_lifecycle::State & configure();
-
-  const rclcpp_lifecycle::State & cleanup();
-
-  const rclcpp_lifecycle::State & shutdown();
-
-  const rclcpp_lifecycle::State & activate();
-
-  const rclcpp_lifecycle::State & deactivate();
-
-  const rclcpp_lifecycle::State & error();
-
-  std::vector<StateInterface::ConstSharedPtr> export_state_interfaces();
-
-  const std::string & get_name() const;
-
-  const std::string & get_group_name() const;
-
-  const rclcpp_lifecycle::State & get_lifecycle_state() const;
-
-  const rclcpp::Time & get_last_read_time() const;
-
-  const HardwareComponentStatisticsCollector & get_read_statistics() const;
-
-  return_type read(const rclcpp::Time & time, const rclcpp::Duration & period);
-
-  return_type write(const rclcpp::Time &, const rclcpp::Duration &) { return return_type::OK; }
-
-  std::recursive_mutex & get_mutex();
-
-private:
-  std::unique_ptr<SensorInterface> impl_;
-  mutable std::recursive_mutex sensors_mutex_;
-  // Last read cycle time
-  rclcpp::Time last_read_cycle_time_;
-  // Component statistics
-  HardwareComponentStatisticsCollector read_statistics_;
-};
-
+using Sensor = HardwareComponent;
 }  // namespace hardware_interface
 #endif  // HARDWARE_INTERFACE__SENSOR_HPP_
